# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * CineQuiz Live API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ServerTime {
  serverTime: number;
}

export interface ErrorResponse {
  error: string;
  code?: string;
}

export interface SuccessResponse {
  success: boolean;
  /** @nullable */
  message?: string | null;
}

export interface AdminCredentials {
  username: string;
  password: string;
}

export interface AuthToken {
  token: string;
  adminId: string;
  username: string;
}

export interface ParticipantRegistration {
  /**
     * @minLength 3
     * @maxLength 20
     */
  rollNumber: string;
  competitionId: string;
}

export interface ParticipantSession {
  participantId: string;
  sessionToken: string;
  competitionId: string;
  rollNumber: string;
}

export interface ParticipantDetail {
  id: string;
  rollNumber: string;
  registeredAt: string;
  currentStatus: string;
  /** @nullable */
  round1Status?: string | null;
  /** @nullable */
  round2Status?: string | null;
  /** @nullable */
  round3Status?: string | null;
  /** @nullable */
  round4Status?: string | null;
}

export interface Competition {
  id: string;
  name: string;
  status: string;
  currentRound: number;
  createdAt: string;
}

export interface RoundPublic {
  id: string;
  roundNumber: number;
  emojiClue: string;
  timeLimitSeconds: number;
  status: string;
  /** @nullable */
  startedAt?: string | null;
  /** @nullable */
  endedAt?: string | null;
}

export interface CompetitionDetail {
  id: string;
  name: string;
  status: string;
  currentRound: number;
  createdAt: string;
  rounds: RoundPublic[];
  participantCount: number;
}

export interface RoundInput {
  competitionId: string;
  roundNumber: number;
  emojiClue: string;
  correctAnswers: string[];
  timeLimitSeconds: number;
}

export interface CompetitionInput {
  name: string;
  rounds?: RoundInput[];
}

export interface CompetitionUpdate {
  name?: string;
  status?: string;
}

export interface Round {
  id: string;
  competitionId: string;
  roundNumber: number;
  emojiClue: string;
  correctAnswers?: string[];
  timeLimitSeconds: number;
  status: string;
  /** @nullable */
  startedAt?: string | null;
  /** @nullable */
  endedAt?: string | null;
}

export interface QualificationDetail {
  id: string;
  participantId: string;
  rollNumber: string;
  roundId: string;
  qualified: boolean;
  /** @nullable */
  qualificationReason?: string | null;
  overrideByAdmin: boolean;
  createdAt: string;
}

export interface QualificationOverride {
  participantId: string;
  roundId: string;
  qualified: boolean;
}

export interface FinalistRevealInput {
  competitionId: string;
}

export interface RoundStat {
  roundNumber: number;
  totalSubmissions: number;
  correctSubmissions: number;
  qualifiedCount: number;
  eliminatedCount: number;
  /** @nullable */
  avgResponseTimeMs?: number | null;
}

export interface FinalistInfo {
  id: string;
  rollNumber: string;
}

export interface CompetitionStats {
  totalParticipants: number;
  roundStats: RoundStat[];
  finalists: FinalistInfo[];
}

export interface AuditLog {
  id: string;
  adminId: string;
  action: string;
  /** @nullable */
  details?: string | null;
  createdAt: string;
}

export type ListParticipantsParams = {
competitionId?: string;
search?: string;
};

export type GetStatsParams = {
competitionId?: string;
};

export type GetAuditLogsParams = {
limit?: number;
};

export type ExportSubmissionsParams = {
competitionId?: string;
};