# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: CineQuiz Live API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /competition:
    get:
      operationId: getCompetition
      tags: [competition]
      summary: Get active competition
      description: Returns current active competition with rounds (answers excluded)
      responses:
        "200":
          description: Competition data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompetitionDetail"
        "404":
          description: No active competition
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createCompetition
      tags: [competition]
      summary: Create competition (admin)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompetitionInput"
      responses:
        "201":
          description: Competition created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompetitionDetail"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /competition/{id}:
    patch:
      operationId: updateCompetition
      tags: [competition]
      summary: Update competition status (admin)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompetitionUpdate"
      responses:
        "200":
          description: Competition updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompetitionDetail"
        "404":
          description: Competition not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /competition/{id}/emergency-stop:
    post:
      operationId: emergencyStop
      tags: [competition]
      summary: Emergency stop (admin)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Competition stopped
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /rounds:
    post:
      operationId: createRound
      tags: [rounds]
      summary: Create or update a round (admin)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoundInput"
      responses:
        "201":
          description: Round created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Round"
  /rounds/{id}:
    get:
      operationId: getRound
      tags: [rounds]
      summary: Get round details (admin)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Round data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Round"
  /rounds/{id}/start:
    post:
      operationId: startRound
      tags: [rounds]
      summary: Start a round (admin)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Round started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Round"
        "400":
          description: Round cannot be started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /rounds/{id}/end:
    post:
      operationId: endRound
      tags: [rounds]
      summary: End a round early (admin)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Round ended
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Round"
...
  /admin/finalist-reveal:
    post:
      operationId: revealFinalists
      tags: [admin]
      summary: Trigger finalist reveal on projector (admin)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FinalistRevealInput"
      responses:
        "200":
          description: Reveal triggered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /admin/stats:
    get:
      operationId: getStats
      tags: [admin]
      summary: Get competition statistics (admin)
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: competitionId
          schema:
            type: string
      responses:
        "200":
          description: Competition statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompetitionStats"
  /admin/audit-logs:
    get:
      operationId: getAuditLogs
      tags: [admin]
      summary: Get audit logs (admin)
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
      responses:
        "200":
          description: Audit logs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AuditLog"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        code:
          type: string
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: ["string", "null"]
      required:
        - success
...
    CompetitionDetail:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        status:
          type: string
        currentRound:
          type: integer
        createdAt:
          type: string
        rounds:
          type: array
          items:
            $ref: "#/components/schemas/RoundPublic"
        participantCount:
          type: integer
      required:
        - id
        - name
        - status
        - currentRound
        - createdAt
        - rounds
        - participantCount
    CompetitionInput:
      type: object
      properties:
        name:
          type: string
        rounds:
          type: array
          items:
            $ref: "#/components/schemas/RoundInput"
      required:
        - name
    CompetitionUpdate:
      type: object
      properties:
        name:
          type: string
        status:
          type: string
    Round:
      type: object
      properties:
        id:
          type: string
        competitionId:
          type: string
        roundNumber:
          type: integer
        emojiClue:
          type: string
        correctAnswers:
          type: array
          items:
            type: string
        timeLimitSeconds:
          type: integer
        status:
          type: string
        startedAt:
          type: ["string", "null"]
        endedAt:
          type: ["string", "null"]
      required:
        - id
        - competitionId
        - roundNumber
        - emojiClue
        - timeLimitSeconds
        - status
    RoundPublic:
      type: object
      properties:
        id:
          type: string
        roundNumber:
          type: integer
        emojiClue:
          type: string
        timeLimitSeconds:
          type: integer
        status:
          type: string
        startedAt:
          type: ["string", "null"]
        endedAt:
          type: ["string", "null"]
      required:
        - id
        - roundNumber
        - emojiClue
        - timeLimitSeconds
        - status
    RoundInput:
      type: object
      properties:
        competitionId:
          type: string
        roundNumber:
          type: integer
        emojiClue:
          type: string
        correctAnswers:
          type: array
          items:
            type: string
        timeLimitSeconds:
          type: integer
      required:
        - competitionId
        - roundNumber
        - emojiClue
        - correctAnswers
        - timeLimitSeconds
...
    FinalistRevealInput:
      type: object
      properties:
        competitionId:
          type: string
      required:
        - competitionId
    CompetitionStats:
      type: object
      properties:
        totalParticipants:
          type: integer
        roundStats:
          type: array
          items:
            $ref: "#/components/schemas/RoundStat"
        finalists:
          type: array
          items:
            $ref: "#/components/schemas/FinalistInfo"
      required:
        - totalParticipants
        - roundStats
        - finalists
    RoundStat:
      type: object
      properties:
        roundNumber:
          type: integer
        totalSubmissions:
          type: integer
        correctSubmissions:
          type: integer
        qualifiedCount:
          type: integer
        eliminatedCount:
          type: integer
        avgResponseTimeMs:
          type: ["integer", "null"]
      required:
        - roundNumber
        - totalSubmissions
        - correctSubmissions
        - qualifiedCount
        - eliminatedCount
    FinalistInfo:
      type: object
      properties:
        id:
          type: string
        rollNumber:
          type: string
      required:
        - id
        - rollNumber
    AuditLog:
      type: object
      properties:
        id:
          type: string
        adminId:
          type: string
        action:
          type: string
        details:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - adminId
        - action
        - createdAt